Make Order a project module instead of a dependency module
so that its external resources can be tracked.
